+++ bank-feeds/src/codatbankfeeds/models/operations/create_connection.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import connection as shared_connection
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import connection as shared_connection
 from codatbankfeeds import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import Optional
@@ -36,8 +35,6 @@ class CreateConnectionResponse:
     r"""HTTP response status code for this operation"""
     connection: Optional[shared_connection.Connection] = dataclasses.field(default=None)
     r"""OK"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""