+++ bank-feeds/src/codatbankfeeds/models/operations/get_bank_account_mapping.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import bankfeedmapping as shared_bankfeedmapping
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import bankfeedmapping as shared_bankfeedmapping
 from typing import Optional
 
 
@@ -26,8 +25,6 @@ class GetBankAccountMappingResponse:
     r"""HTTP response status code for this operation"""
     bank_feed_mapping: Optional[shared_bankfeedmapping.BankFeedMapping] = dataclasses.field(default=None)
     r"""Success"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""