+++ bank-feeds/RELEASES.md
@@ -739,3 +739,13 @@ Based on:
 - [python v4.1.0] bank-feeds
 ### Releases
 - [PyPI v4.1.0] https://pypi.org/project/codat-bankfeeds/4.1.0 - bank-feeds
+
+## 2023-11-13 01:28:36
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v5.0.0] bank-feeds
+### Releases
+- [PyPI v5.0.0] https://pypi.org/project/codat-bankfeeds/5.0.0 - bank-feeds